Tersify summary
I think the summary is to much describing what a business case is and benefits from cutting to the chase.

I debated if the four areas of benefits should be linked, but then that would be even more redundant with the Page Contents block, so I didn’t do that.

Signed-off-by: Eric Eggert <mail@yatil.net>