Who needs a business case?
Changed the lead in for the section a little bit, because I think not only the term but the concept of a business case is important.

I removed the use of ”We” as we usually don’t do that and it was only done once in the article.

Minor sentence simplifications/clarifications.

Signed-off-by: Eric Eggert <mail@yatil.net>